use ints not shorts
git-svn-id: https://nocs.stanford.edu/cgi-bin/svn.cgi/booksim2.0/trunk@2811 ce7aba0d-9f81-40a6-a495-cecae84b0515

+++ src/trafficmanager.hpp
@@ -87,7 +87,7 @@ protected:
 
   // ============ sub-networks and deadlock ==========
 
-  short _duplicate_networks;
+  int _duplicate_networks;
   int _deadlock_timer;
   int _deadlock_warn_timeout;
 
@@ -188,8 +188,8 @@ protected:
   int   _sample_period;
   int   _max_samples;
   int   _warmup_periods;
-  vector<vector<short> > _class_array;
-  short _sub_network;
+  vector<vector<int> > _class_array;
+  int   _sub_network;
 
   int   _include_queuing;
